:truck: move functions to module
- :bug: some function still hidden in 1_batch_analysis page
